mod: address PR review — rename ALWAYS_ENDPOINTS, extract multipart parser
- Rename ALWAYS_ENDPOINTS to MANAGEMENT_ENDPOINTS for clarity
- Extract multipart/form-data parsing from http.ts into multipart.ts

Made-with: Cursor
